Added deletion of join from point and oevent ids Started with test for database
