+++ TRing/app/build.gradle
@@ -59,6 +59,7 @@ dependencies {
     //RxJava
     compile 'io.reactivex.rxjava2:rxjava:2.0.6'
     compile 'io.reactivex.rxjava2:rxandroid:2.0.1'
+    androidTestImplementation "android.arch.core:core-testing:1.0.0-alpha1"
 
     //Test migrations
     testCompile 'junit:junit:4.12'
