+++ TRing/app/src/main/java/no/teacherspet/tring/Database/Entities/PointOEventJoin.java
@@ -3,7 +3,6 @@ package no.teacherspet.tring.Database.Entities;
 import android.arch.persistence.room.Entity;
 import android.arch.persistence.room.ForeignKey;
 import android.arch.persistence.room.Index;
-import android.arch.persistence.room.OnConflictStrategy;
 
 /**
  * Created by Hermann on 15.02.2018.
